+++ beacon-chain/operations/blstoexec/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "doc.go",
         "pool.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/v3/beacon-chain/operations/blstoexecchanges",
+    importpath = "github.com/prysmaticlabs/prysm/v3/beacon-chain/operations/blstoexec",
     visibility = [
         "//beacon-chain:__subpackages__",
     ],
