+++ beacon-chain/operations/blstoexec/doc.go
@@ -2,4 +2,4 @@
 // BLS to execution change events by the beacon node, handling their lifecycle
 // and performing integrity checks before serving them as objects
 // for validators to include in blocks.
-package blstoexecchanges
+package blstoexec
