+++ beacon-chain/operations/blstoexec/pool.go
@@ -1,4 +1,4 @@
-package blstoexecchanges
+package blstoexec
 
 import (
 	"sort"
