+++ beacon-chain/operations/blstoexec/pool_test.go
@@ -1,4 +1,4 @@
-package blstoexecchanges
+package blstoexec
 
 import (
 	"sync"
